requirements.txt
boto3>=1.26.0
requests>=2.28.0
pyyaml>=6.0
